+++ Changelog.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- None
+- Digital rock Module
+    - Region module
+        - Sphere class
+        - Cylinder class
 
 ### Changed
 - None